Spread people
I made the JavaScript work. I got a hint to put the JavaScript file at the end
of the HTML and now it works lika a charm.
I wrappte the JavaScript conntetion between the HTML and Ruby away in the Panels
class. So I have only one presentation of the gamefield.